+++ CHANGELOG.md
@@ -4,7 +4,20 @@ This changelog follows the specifications detailed in: [Keep a Changelog](https:
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
 
-## Version 0.7.6 - Unreleased
+## Version 0.7.7 - Unreleased
+
+### Added
+
+* `Boxes.to_slices`
+* `Affine.concise`
+
+
+### Fixed
+
+* counterclockwise polygon check
+
+
+## Version 0.7.6 - Released 2021-05-25
 
 ### Fixed